+++ lib/contentful/management/organization.rb
@@ -15,6 +15,22 @@ module Contentful
       def self.build_endpoint(_endpoint_options)
         'organizations'
       end
+
+      # Allows listing all usage periods for organization.
+      # @see _ README for details.
+      #
+      # @return [Contentful::Management::ClientApiUsageMethodsFactory]
+      def usage_periods
+        ClientApiUsageMethodsFactory.new(client, id)
+      end
+
+      # Allows listing all api usage for organization.
+      # @see _ README for details.
+      #
+      # @return [Contentful::Management::ClientApiUsageMethodsFactory]
+      def api_usage
+        ClientApiUsageMethodsFactory.new(client, id)
+      end
     end
   end
 end